pyproject.toml
[project]
name = "email-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "django>=5.2.4",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "geocoder>=1.38.1",
    "google>=3.0.0",
    "google-api-python-client>=2.178.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "osmnx>=2.0.5",
    "overpy>=0.7",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "thefuck>=3.32",
]
